+++ sdk/resourcemanager/sqlvirtualmachine/armsqlvirtualmachine/constants.go
@@ -11,7 +11,7 @@ package armsqlvirtualmachine
 
 const (
 	moduleName    = "armsqlvirtualmachine"
-	moduleVersion = "v0.7.0"
+	moduleVersion = "v0.8.0"
 )
 
 // AssessmentDayOfWeek - Day of the week to run assessment.
@@ -460,5 +460,19 @@ func PossibleStorageWorkloadTypeValues() []StorageWorkloadType {
 		StorageWorkloadTypeDW,
 		StorageWorkloadTypeGENERAL,
 		StorageWorkloadTypeOLTP,
+	}
+}
+
+// TroubleShootingScenario - SQL VM troubleshooting scenario.
+type TroubleShootingScenario string
+
+const (
+	TroubleShootingScenarioUnHealthyReplica TroubleShootingScenario = "UnHealthyReplica"
+)
+
+// PossibleTroubleShootingScenarioValues returns the possible values for the TroubleShootingScenario const type.
+func PossibleTroubleShootingScenarioValues() []TroubleShootingScenario {
+	return []TroubleShootingScenario{
+		TroubleShootingScenarioUnHealthyReplica,
 	}
 }