+++ sdk/resourcemanager/sqlvirtualmachine/armsqlvirtualmachine/response_types.go
@@ -108,3 +108,8 @@ type SQLVirtualMachinesClientStartAssessmentResponse struct {
 type SQLVirtualMachinesClientUpdateResponse struct {
 	SQLVirtualMachine
 }
+
+// TroubleshootClientTroubleshootResponse contains the response from method TroubleshootClient.Troubleshoot.
+type TroubleshootClientTroubleshootResponse struct {
+	// placeholder for future response values
+}